pyproject.toml
[tool.poetry]
name = "owlreads"
version = "0.1.0"
description = ""
authors = ["Neha Kerung <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.2.8"
djangorestframework = "^3.16.1"
psycopg2 = "^2.9.11"
django-filter = "^25.2"
django-split-settings = "^1.3.2"
pyyaml = "^6.0.3"
python-decouple = "^3.8"
requests = "^2.32.5"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.3.0"
pre-commit = "^4.4.0"

[tool.isort]
multi_line_output = 5
line_length = 120

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
